Remove UrlValidator dependency on util/Flags and switch to using "long" for flags for URL and CreditCard options
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/validator/trunk@594937 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/validator/routines/CreditCardValidator.java
@@ -54,32 +54,32 @@ public class CreditCardValidator implements Serializable {
      * v.isValid(aCardNumber);
      * </pre>
      */
-    public static final int NONE = 0;
+    public static final long NONE = 0;
 
     /**
      * Option specifying that American Express cards are allowed.
      */
-    public static final int AMEX = 1 << 0;
+    public static final long AMEX = 1 << 0;
 
     /**
      * Option specifying that Visa cards are allowed.
      */
-    public static final int VISA = 1 << 1;
+    public static final long VISA = 1 << 1;
 
     /**
      * Option specifying that Mastercard cards are allowed.
      */
-    public static final int MASTERCARD = 1 << 2;
+    public static final long MASTERCARD = 1 << 2;
 
     /**
      * Option specifying that Discover cards are allowed.
      */
-    public static final int DISCOVER = 1 << 3;
+    public static final long DISCOVER = 1 << 3;
 
     /**
      * Option specifying that Diners cards are allowed.
      */
-    public static final int DINERS = 1 << 4;
+    public static final long DINERS = 1 << 4;
     
     /**
      * The CreditCardTypes that are allowed to pass validation.
@@ -122,7 +122,7 @@ public class CreditCardValidator implements Serializable {
      * CreditCardValidator.VISA + CreditCardValidator.AMEX to specify that 
      * those are the only valid card types.
      */
-    public CreditCardValidator(int options) {
+    public CreditCardValidator(long options) {
         super();
 
         if (isOn(options, VISA)) {
@@ -207,7 +207,7 @@ public class CreditCardValidator implements Serializable {
      *
      * @return whether the specified flag value is on.
      */
-    private boolean isOn(int options, int flag) {
+    private boolean isOn(long options, long flag) {
         return (options & flag) > 0;
     }
